requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.5
filelock==3.12.2
fonttools==4.40.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.5.post0
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.117
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
